Stop the beds cast eating a half bedroom
The test expected '3' and got '3.0', which looked like a test problem and
was not. leads.beds is numeric(4,1) — the column exists at that precision
so a 3.5-bedroom house can be recorded — and the enrich trigger cast the
payload value to ::int on the way in. Every half bedroom a vendor sent
would have been silently rounded down.

baths never had the cast, so the two columns disagreed about the same
kind of value.

Cast removed, test expectation corrected to '3.0', and an assertion added
that "3.5" survives as 3.5 rather than becoming 3.

+++ supabase/migrations/20260902160000_enrichment_can_never_reject_a_lead.sql
@@ -88,7 +88,10 @@ BEGIN
     NEW.county        := coalesce(NEW.county,        public.payload_value(f, 'county'));
     NEW.property_type := coalesce(NEW.property_type, public.payload_value(f, 'property_type'));
 
-    NEW.beds  := coalesce(NEW.beds,  public.payload_number(f, 'beds')::int);
+    -- beds is numeric(4,1), not an integer. The ::int cast this used to carry
+    -- would have quietly turned a 3.5-bedroom house into a 3-bedroom one —
+    -- found because the test expected '3' and the column returned '3.0'.
+    NEW.beds  := coalesce(NEW.beds,  public.payload_number(f, 'beds'));
     NEW.baths := coalesce(NEW.baths, public.payload_number(f, 'baths'));
     NEW.sqft  := coalesce(NEW.sqft,  public.payload_number(f, 'sqft')::int);
     NEW.year_built := coalesce(NEW.year_built,
